monster parsing: move spell functions
